PR review updates: renamed s3-related modules: atum-sdk-s3 -> atum-s3-sdk-extension, atum-examples-sdk-s3 -> atum-examples-s3-sdk-extension; Absa header added where missing corrected package names za.co.absa.atum.persistence.*
