+++ atum-s3-sdk-extension/src/main/scala/za/co/absa/atum/core/AtumSdkS3.scala
