+++ atum-s3-sdk-extension/src/main/scala/za/co/absa/atum/core/ControlFrameworkStateSdkS3.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2018-2019 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.atum.core
+
+import org.apache.spark.sql.SparkSession
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider
+import za.co.absa.atum.persistence.s3.{ControlMeasuresSdkS3StorerJsonFile, S3KmsSettings, SimpleS3LocationWithRegion}
+
+/**
+  * This class holds control measurement's state
+  */
+class ControlFrameworkStateSdkS3(sparkSession: SparkSession) extends ControlFrameworkState(sparkSession) {
+
+  def storeCurrentInfoFileOnSdkS3(s3Location: SimpleS3LocationWithRegion, s3KmsSettings: S3KmsSettings)(implicit credentialsProvider: AwsCredentialsProvider): Unit = {
+    val storer = ControlMeasuresSdkS3StorerJsonFile(s3Location, s3KmsSettings)
+
+    storer.store(accumulator.getControlMeasure)
+    AtumSdkS3.log.info(s"Control measurements saved to ${s3Location.s3String}")
+  }
+}
