+++ atum-s3-sdk-extension/src/test/resources/example_input.info
