+++ atum-s3-sdk-extension/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
