--- atum-sdk-s3/src/main/scala/za/co/absa/atum/persistence/S3ControlMeasuresStorer.scala
@@ -1,24 +1,0 @@
-/*
- * Copyright 2018-2019 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.atum.persistence
-
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider
-
-trait S3ControlMeasuresStorer extends ControlMeasuresStorer {
-  def kmsSettings: S3KmsSettings
-  def outputLocation: SimpleS3LocationWithRegion
-  def credentialsProvider: AwsCredentialsProvider
-}
