+++ atum/src/main/scala/za/co/absa/atum/utils/BuildProperties.scala
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
 package za.co.absa.atum.utils
