+++ atum/src/main/scala/za/co/absa/atum/utils/S3LocationUtils.scala
@@ -1,15 +1,32 @@
+/*
+ * Copyright 2018-2019 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package za.co.absa.atum.utils
+
+import scala.util.matching.Regex
 
 object S3LocationUtils {
 
   // hint: https://docs.aws.amazon.com/AmazonS3/latest/dev/BucketRestrictions.html#bucketnamingrules
-  val S3LocationRx = "(s3[an]?)://([-a-z0-9.]{3,63})/(.*)".r
+  val S3LocationRx: Regex = "(s3[an]?)://([-a-z0-9.]{3,63})/(.*)".r
 
   implicit class StringS3LocationExt(path: String) {
 
     def toS3Location: Option[SimpleS3Location] = {
       path match {
-        case S3LocationRx(protocol, bucketName, path) => Some(SimpleS3Location(protocol, bucketName, path))
+        case S3LocationRx(protocol, bucketName, relativePath) => Some(SimpleS3Location(protocol, bucketName, relativePath))
         case _ => None
       }
     }
