--- examples-sdk-s3/pom.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright 2018-2019 ABSA Group Limited
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>atum-examples-sdk-s3_2.11</artifactId>
-    <packaging>jar</packaging>
-
-    <parent>
-        <groupId>za.co.absa</groupId>
-        <artifactId>atum-parent_2.11</artifactId>
-        <version>3.2.1-SNAPSHOT</version>
-    </parent>
-
-    <dependencies>
-        <!-- Atum -->
-        <dependency>
-			<groupId>za.co.absa</groupId>
-			<artifactId>atum-sdk-s3_${scala.binary.version}</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-        <dependency>
-            <groupId>za.co.absa</groupId>
-            <artifactId>atum-examples_${scala.binary.version}</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <directory>${project.basedir}/target/${scala.binary.version}</directory>
-        <outputDirectory>${project.build.directory}/classes</outputDirectory>
-        <testOutputDirectory>${project.build.directory}/test-classes</testOutputDirectory>
-
-        <sourceDirectory>src/main/scala</sourceDirectory>
-        <testSourceDirectory>src/test/scala</testSourceDirectory>
-        <plugins>
-            <plugin>
-                <groupId>org.scalatest</groupId>
-                <artifactId>scalatest-maven-plugin</artifactId>
-                <version>${scalatest.maven.version}</version>
-                <configuration>
-                    <skipTests>true</skipTests>
-                </configuration>
-            </plugin>
-            <!-- Uber jar generation -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>${maven.shade.version}</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
-    <profiles>
-        <profile>
-            <id>scala-2.11</id>
-            <properties>
-                <scala.binary.version>2.11</scala.binary.version>
-                <scala.version>${scala_2.11.version}</scala.version>
-            </properties>
-        </profile>
-        <profile>
-            <id>scala-2.12</id>
-            <properties>
-                <scala.binary.version>2.12</scala.binary.version>
-                <scala.version>${scala_2.12.version}</scala.version>
-            </properties>
-        </profile>
-    </profiles>
-</project>
