--- examples-sdk-s3/src/test/scala/za/co/absa/atum/examples/SampleMeasurementsSdkS3RunnerSpec.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2018-2019 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.atum.examples
-
-import org.scalatest.Ignore
-import org.scalatest.funsuite.AnyFunSuite
-import za.co.absa.atum.utils._
-
-// integ tests are skipped via pom-settings
-class SampleMeasurementsS3RunnerExampleSpec extends AnyFunSuite
-  with SparkJobRunnerMethods
-  with SparkLocalMaster {
-
-    runSparkJobAsTest[SampleSdkS3Measurements1.type]
-    runSparkJobAsTest[SampleSdkS3Measurements2.type]
-}
